Prevent tobira tab crash when ID is missing
I recently observed that in some cases, the Tobira
API does not send an ID for its respective asset
(i.e. video or series).
I have not been able to reproduce this anymore, so
it's hard to identify the unterlying issue. But since
others have reported this as well, in order to prevent
this leading to crashes, this now uses optional chaining
for `id` operations.

This won't break anything, it just means that in cases
where the (Tobira) ID is actually missing, the Opencast
ID is used in links to these assets instead.