Switch to react-tooltips
Replaces our mui tooltip wrapper with react-tooltip. Also adds back
tooltips that were commented out.

The reason for this change is that tooltips from react-tooltip are
much more performant
when used in larger quantities. It will still add about ~100ms to
the render time of events or series
tables with a 100 elements, which is not great but hardly noticable
with our load times of over a second.